requirements.txt
asgiref==3.8.1
Django==4.2.6
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==10.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2